+++ DemandaFacil/src/views/cliente/TelaLogin.form
@@ -300,10 +300,8 @@
                                           <EmptySpace pref="63" max="32767" attributes="0"/>
                                           <Component id="jButton1" min="-2" max="-2" attributes="0"/>
                                       </Group>
-                                      <Group type="103" groupAlignment="0" max="-2" attributes="0">
-                                          <Component id="jTextField1" max="32767" attributes="0"/>
-                                          <Component id="jPasswordField1" pref="171" max="32767" attributes="0"/>
-                                      </Group>
+                                      <Component id="jTextField1" alignment="0" max="32767" attributes="0"/>
+                                      <Component id="jPasswordField1" alignment="0" pref="171" max="32767" attributes="0"/>
                                   </Group>
                                   <EmptySpace min="0" pref="55" max="32767" attributes="0"/>
                               </Group>
