--- DemandaFacil/src/views/produtos/Cadastro_Cliente.java
@@ -1,566 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package views.empresa;
-
-/**
- *
- * @author souza
- */
-public class Cadastro_Cliente extends javax.swing.JFrame {
-
-    /**
-     * Creates new form Perfil_Empresa
-     */
-    public Cadastro_Cliente() {
-        initComponents();
-    }
-
-    /**
-     * This method is called from within the constructor to initialize the form.
-     * WARNING: Do NOT modify this code. The content of this method is always
-     * regenerated by the Form Editor.
-     */
-    @SuppressWarnings("unchecked")
-    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
-    private void initComponents() {
-
-        jPanel1 = new javax.swing.JPanel();
-        background = new javax.swing.JPanel();
-        faixaTitulo = new javax.swing.JPanel();
-        subTitulo = new javax.swing.JLabel();
-        nomePagina = new javax.swing.JLabel();
-        menuLateral = new javax.swing.JPanel();
-        linhaDivisoria = new javax.swing.JPanel();
-        opcaoMenu = new javax.swing.JPanel();
-        icon = new javax.swing.JLabel();
-        nome = new javax.swing.JLabel();
-        opcaoMenu1 = new javax.swing.JPanel();
-        icon1 = new javax.swing.JLabel();
-        nome1 = new javax.swing.JLabel();
-        opcaoMenu2 = new javax.swing.JPanel();
-        icon2 = new javax.swing.JLabel();
-        nome2 = new javax.swing.JLabel();
-        opcaoMenu3 = new javax.swing.JPanel();
-        icon3 = new javax.swing.JLabel();
-        nome3 = new javax.swing.JLabel();
-        iconLogo = new javax.swing.JLabel();
-        jLabel13 = new javax.swing.JLabel();
-        txt_perfilEmpresa2 = new javax.swing.JTextField();
-        txt_perfilEmpresa6 = new javax.swing.JTextField();
-        painelBackgroundForm = new javax.swing.JPanel();
-        painelForm = new javax.swing.JPanel();
-        jLabel5 = new javax.swing.JLabel();
-        txt_perfilEmpresa = new javax.swing.JTextField();
-        jLabel6 = new javax.swing.JLabel();
-        jLabel8 = new javax.swing.JLabel();
-        txt_perfilEmpresa1 = new javax.swing.JTextField();
-        txt_perfilEmpresa7 = new javax.swing.JTextField();
-        jLabel7 = new javax.swing.JLabel();
-        txt_perfilEmpresa4 = new javax.swing.JTextField();
-        jLabel1 = new javax.swing.JLabel();
-        txt_perfilEmpresa5 = new javax.swing.JTextField();
-        jLabel9 = new javax.swing.JLabel();
-        txt_perfilEmpresa8 = new javax.swing.JTextField();
-        jLabel11 = new javax.swing.JLabel();
-        txt_perfilEmpresa11 = new javax.swing.JTextField();
-        jLabel14 = new javax.swing.JLabel();
-        jLabel3 = new javax.swing.JLabel();
-        jTextField1 = new javax.swing.JTextField();
-        txt_perfilEmpresa13 = new javax.swing.JTextField();
-        jLabel15 = new javax.swing.JLabel();
-        txt_perfilEmpresa12 = new javax.swing.JTextField();
-        jLabel16 = new javax.swing.JLabel();
-        txt_perfilEmpresa15 = new javax.swing.JTextField();
-        jButton2 = new javax.swing.JButton();
-        jLabel4 = new javax.swing.JLabel();
-        txt_perfilEmpresa16 = new javax.swing.JTextField();
-
-        setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
-
-        background.setBackground(new java.awt.Color(178, 242, 236));
-        background.setLayout(new org.netbeans.lib.awtextra.AbsoluteLayout());
-
-        faixaTitulo.setBackground(new java.awt.Color(0, 128, 128));
-
-        subTitulo.setFont(new java.awt.Font("Painted Paradise", 0, 24)); // NOI18N
-        subTitulo.setForeground(java.awt.Color.white);
-        subTitulo.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        subTitulo.setText("Cadastro");
-        subTitulo.setMaximumSize(new java.awt.Dimension(90, 24));
-        subTitulo.setMinimumSize(new java.awt.Dimension(90, 24));
-
-        nomePagina.setFont(new java.awt.Font("Painted Paradise", 1, 36)); // NOI18N
-        nomePagina.setForeground(java.awt.Color.white);
-        nomePagina.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        nomePagina.setText("Cliente");
-
-        javax.swing.GroupLayout faixaTituloLayout = new javax.swing.GroupLayout(faixaTitulo);
-        faixaTitulo.setLayout(faixaTituloLayout);
-        faixaTituloLayout.setHorizontalGroup(
-            faixaTituloLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(faixaTituloLayout.createSequentialGroup()
-                .addGroup(faixaTituloLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(faixaTituloLayout.createSequentialGroup()
-                        .addGap(47, 47, 47)
-                        .addComponent(subTitulo, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                    .addGroup(faixaTituloLayout.createSequentialGroup()
-                        .addContainerGap()
-                        .addComponent(nomePagina, javax.swing.GroupLayout.PREFERRED_SIZE, 277, javax.swing.GroupLayout.PREFERRED_SIZE)))
-                .addContainerGap(283, Short.MAX_VALUE))
-        );
-        faixaTituloLayout.setVerticalGroup(
-            faixaTituloLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(faixaTituloLayout.createSequentialGroup()
-                .addGap(18, 18, 18)
-                .addComponent(subTitulo, javax.swing.GroupLayout.PREFERRED_SIZE, 32, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(nomePagina, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                .addContainerGap())
-        );
-
-        background.add(faixaTitulo, new org.netbeans.lib.awtextra.AbsoluteConstraints(230, 0, 570, -1));
-
-        menuLateral.setBackground(new java.awt.Color(47, 79, 79));
-
-        linhaDivisoria.setBackground(new java.awt.Color(80, 183, 173));
-
-        javax.swing.GroupLayout linhaDivisoriaLayout = new javax.swing.GroupLayout(linhaDivisoria);
-        linhaDivisoria.setLayout(linhaDivisoriaLayout);
-        linhaDivisoriaLayout.setHorizontalGroup(
-            linhaDivisoriaLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 0, Short.MAX_VALUE)
-        );
-        linhaDivisoriaLayout.setVerticalGroup(
-            linhaDivisoriaLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 8, Short.MAX_VALUE)
-        );
-
-        opcaoMenu.setBackground(new java.awt.Color(57, 95, 95));
-        opcaoMenu.setLayout(new javax.swing.BoxLayout(opcaoMenu, javax.swing.BoxLayout.LINE_AXIS));
-
-        icon.setIcon(new javax.swing.ImageIcon(getClass().getResource("/views/imagens/icones/icons8-Home.png"))); // NOI18N
-        opcaoMenu.add(icon);
-
-        nome.setFont(new java.awt.Font("Merriweather Light", 1, 24)); // NOI18N
-        nome.setForeground(java.awt.Color.white);
-        nome.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        nome.setText("Home");
-        nome.setMaximumSize(new java.awt.Dimension(80, 24));
-        nome.setMinimumSize(new java.awt.Dimension(58, 24));
-        opcaoMenu.add(nome);
-
-        opcaoMenu1.setBackground(new java.awt.Color(57, 95, 95));
-        opcaoMenu1.setLayout(new javax.swing.BoxLayout(opcaoMenu1, javax.swing.BoxLayout.LINE_AXIS));
-
-        icon1.setIcon(new javax.swing.ImageIcon(getClass().getResource("/views/imagens/icones/icons8-Exit.png"))); // NOI18N
-        opcaoMenu1.add(icon1);
-
-        nome1.setFont(new java.awt.Font("Merriweather Light", 1, 24)); // NOI18N
-        nome1.setForeground(java.awt.Color.white);
-        nome1.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        nome1.setText("Sair");
-        nome1.setMaximumSize(new java.awt.Dimension(70, 24));
-        nome1.setMinimumSize(new java.awt.Dimension(46, 24));
-        opcaoMenu1.add(nome1);
-
-        opcaoMenu2.setBackground(new java.awt.Color(57, 95, 95));
-        opcaoMenu2.setLayout(new javax.swing.BoxLayout(opcaoMenu2, javax.swing.BoxLayout.LINE_AXIS));
-
-        icon2.setIcon(new javax.swing.ImageIcon(getClass().getResource("/views/imagens/icones/icons8-Business.png"))); // NOI18N
-        opcaoMenu2.add(icon2);
-
-        nome2.setFont(new java.awt.Font("Merriweather Light", 1, 24)); // NOI18N
-        nome2.setForeground(java.awt.Color.white);
-        nome2.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        nome2.setText("Empresas");
-        nome2.setMaximumSize(new java.awt.Dimension(130, 32));
-        nome2.setMinimumSize(new java.awt.Dimension(95, 24));
-        opcaoMenu2.add(nome2);
-
-        opcaoMenu3.setBackground(new java.awt.Color(57, 95, 95));
-        opcaoMenu3.setLayout(new javax.swing.BoxLayout(opcaoMenu3, javax.swing.BoxLayout.LINE_AXIS));
-
-        icon3.setIcon(new javax.swing.ImageIcon(getClass().getResource("/views/imagens/icones/icons8-More Info.png"))); // NOI18N
-        opcaoMenu3.add(icon3);
-
-        nome3.setFont(new java.awt.Font("Merriweather Light", 1, 24)); // NOI18N
-        nome3.setForeground(java.awt.Color.white);
-        nome3.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        nome3.setText("Perfil");
-        nome3.setMaximumSize(new java.awt.Dimension(80, 24));
-        nome3.setMinimumSize(new java.awt.Dimension(65, 24));
-        opcaoMenu3.add(nome3);
-
-        iconLogo.setFont(new java.awt.Font("Ubuntu", 0, 24)); // NOI18N
-        iconLogo.setForeground(java.awt.Color.white);
-        iconLogo.setText("Logo");
-
-        javax.swing.GroupLayout menuLateralLayout = new javax.swing.GroupLayout(menuLateral);
-        menuLateral.setLayout(menuLateralLayout);
-        menuLateralLayout.setHorizontalGroup(
-            menuLateralLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addComponent(linhaDivisoria, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-            .addComponent(opcaoMenu, javax.swing.GroupLayout.DEFAULT_SIZE, 230, Short.MAX_VALUE)
-            .addComponent(opcaoMenu2, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-            .addComponent(opcaoMenu1, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-            .addComponent(opcaoMenu3, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-            .addGroup(menuLateralLayout.createSequentialGroup()
-                .addGap(82, 82, 82)
-                .addComponent(iconLogo)
-                .addContainerGap(95, Short.MAX_VALUE))
-        );
-        menuLateralLayout.setVerticalGroup(
-            menuLateralLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(menuLateralLayout.createSequentialGroup()
-                .addGap(42, 42, 42)
-                .addComponent(iconLogo)
-                .addGap(40, 40, 40)
-                .addComponent(linhaDivisoria, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(50, 50, 50)
-                .addComponent(opcaoMenu, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(opcaoMenu3, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(opcaoMenu2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                .addComponent(opcaoMenu1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(38, 38, 38))
-        );
-
-        background.add(menuLateral, new org.netbeans.lib.awtextra.AbsoluteConstraints(0, 0, 230, 540));
-
-        jLabel13.setFont(new java.awt.Font("Tahoma", 0, 12)); // NOI18N
-        jLabel13.setText("Bairro:");
-        background.add(jLabel13, new org.netbeans.lib.awtextra.AbsoluteConstraints(0, 0, -1, -1));
-        background.add(txt_perfilEmpresa2, new org.netbeans.lib.awtextra.AbsoluteConstraints(0, 0, -1, -1));
-
-        txt_perfilEmpresa6.setText("         /        /");
-        txt_perfilEmpresa6.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                txt_perfilEmpresa6ActionPerformed(evt);
-            }
-        });
-        background.add(txt_perfilEmpresa6, new org.netbeans.lib.awtextra.AbsoluteConstraints(0, 0, -1, -1));
-
-        painelBackgroundForm.setBackground(java.awt.Color.lightGray);
-
-        jLabel5.setFont(new java.awt.Font("Tahoma", 0, 12)); // NOI18N
-        jLabel5.setText("Nome:");
-
-        jLabel6.setFont(new java.awt.Font("Tahoma", 0, 12)); // NOI18N
-        jLabel6.setText("CPF:");
-
-        jLabel8.setFont(new java.awt.Font("Tahoma", 0, 12)); // NOI18N
-        jLabel8.setText("RG:");
-
-        jLabel7.setFont(new java.awt.Font("Tahoma", 0, 12)); // NOI18N
-        jLabel7.setText("E-mail:");
-
-        jLabel1.setFont(new java.awt.Font("Tahoma", 0, 12)); // NOI18N
-        jLabel1.setText("Data de Nascimento:");
-
-        txt_perfilEmpresa5.setText("         /        /");
-        txt_perfilEmpresa5.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                txt_perfilEmpresa5ActionPerformed(evt);
-            }
-        });
-
-        jLabel9.setFont(new java.awt.Font("Tahoma", 0, 12)); // NOI18N
-        jLabel9.setText("Cidade:");
-
-        jLabel11.setFont(new java.awt.Font("Tahoma", 0, 12)); // NOI18N
-        jLabel11.setText("Endereço:");
-
-        jLabel14.setFont(new java.awt.Font("Tahoma", 0, 12)); // NOI18N
-        jLabel14.setText("Telefone:      ");
-
-        jLabel3.setFont(new java.awt.Font("Tahoma", 0, 12)); // NOI18N
-        jLabel3.setText("Usuário:");
-
-        jTextField1.setText("(   )");
-
-        jLabel15.setFont(new java.awt.Font("Tahoma", 0, 12)); // NOI18N
-        jLabel15.setText("Estado:");
-
-        jLabel16.setFont(new java.awt.Font("Tahoma", 0, 12)); // NOI18N
-        jLabel16.setText("Bairro:");
-
-        jButton2.setText("Confirmar");
-
-        jLabel4.setFont(new java.awt.Font("Tahoma", 0, 12)); // NOI18N
-        jLabel4.setText("Senha:");
-
-        javax.swing.GroupLayout painelFormLayout = new javax.swing.GroupLayout(painelForm);
-        painelForm.setLayout(painelFormLayout);
-        painelFormLayout.setHorizontalGroup(
-            painelFormLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(painelFormLayout.createSequentialGroup()
-                .addContainerGap()
-                .addGroup(painelFormLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(painelFormLayout.createSequentialGroup()
-                        .addGroup(painelFormLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                            .addGroup(painelFormLayout.createSequentialGroup()
-                                .addComponent(jLabel5)
-                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                                .addComponent(txt_perfilEmpresa, javax.swing.GroupLayout.PREFERRED_SIZE, 210, javax.swing.GroupLayout.PREFERRED_SIZE))
-                            .addGroup(painelFormLayout.createSequentialGroup()
-                                .addGroup(painelFormLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                                    .addComponent(jLabel6)
-                                    .addComponent(jLabel8))
-                                .addGap(18, 18, 18)
-                                .addGroup(painelFormLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                                    .addComponent(txt_perfilEmpresa7, javax.swing.GroupLayout.PREFERRED_SIZE, 118, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                    .addComponent(txt_perfilEmpresa1, javax.swing.GroupLayout.PREFERRED_SIZE, 118, javax.swing.GroupLayout.PREFERRED_SIZE)))
-                            .addGroup(painelFormLayout.createSequentialGroup()
-                                .addComponent(jLabel7)
-                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                                .addComponent(txt_perfilEmpresa4, javax.swing.GroupLayout.PREFERRED_SIZE, 210, javax.swing.GroupLayout.PREFERRED_SIZE)))
-                        .addGap(550, 550, 550))
-                    .addGroup(painelFormLayout.createSequentialGroup()
-                        .addGroup(painelFormLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                            .addComponent(jLabel14)
-                            .addGroup(painelFormLayout.createSequentialGroup()
-                                .addGroup(painelFormLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
-                                    .addComponent(jLabel9, javax.swing.GroupLayout.Alignment.LEADING)
-                                    .addGroup(painelFormLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                                        .addGroup(painelFormLayout.createSequentialGroup()
-                                            .addComponent(jLabel1)
-                                            .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                                            .addComponent(txt_perfilEmpresa5, javax.swing.GroupLayout.PREFERRED_SIZE, 86, javax.swing.GroupLayout.PREFERRED_SIZE))
-                                        .addGroup(painelFormLayout.createSequentialGroup()
-                                            .addGroup(painelFormLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                                                .addComponent(jLabel11)
-                                                .addComponent(jLabel3))
-                                            .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                                            .addGroup(painelFormLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                                                .addComponent(jTextField1, javax.swing.GroupLayout.PREFERRED_SIZE, 120, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                                .addComponent(txt_perfilEmpresa11, javax.swing.GroupLayout.PREFERRED_SIZE, 196, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                                .addComponent(txt_perfilEmpresa13, javax.swing.GroupLayout.PREFERRED_SIZE, 120, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                                .addComponent(txt_perfilEmpresa8, javax.swing.GroupLayout.PREFERRED_SIZE, 196, javax.swing.GroupLayout.PREFERRED_SIZE)))))
-                                .addGap(35, 35, 35)
-                                .addGroup(painelFormLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                                    .addGroup(painelFormLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
-                                        .addGroup(painelFormLayout.createSequentialGroup()
-                                            .addComponent(jLabel15)
-                                            .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                                            .addComponent(txt_perfilEmpresa12))
-                                        .addGroup(painelFormLayout.createSequentialGroup()
-                                            .addComponent(jLabel16)
-                                            .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                                            .addComponent(txt_perfilEmpresa15, javax.swing.GroupLayout.PREFERRED_SIZE, 120, javax.swing.GroupLayout.PREFERRED_SIZE)))
-                                    .addGroup(painelFormLayout.createSequentialGroup()
-                                        .addComponent(jLabel4)
-                                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                                        .addComponent(txt_perfilEmpresa16, javax.swing.GroupLayout.PREFERRED_SIZE, 122, javax.swing.GroupLayout.PREFERRED_SIZE)))))
-                        .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))))
-            .addGroup(painelFormLayout.createSequentialGroup()
-                .addGap(217, 217, 217)
-                .addComponent(jButton2)
-                .addGap(0, 0, Short.MAX_VALUE))
-        );
-        painelFormLayout.setVerticalGroup(
-            painelFormLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(painelFormLayout.createSequentialGroup()
-                .addGap(35, 35, 35)
-                .addGroup(painelFormLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(jLabel5)
-                    .addComponent(txt_perfilEmpresa, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addGap(11, 11, 11)
-                .addGroup(painelFormLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(jLabel6)
-                    .addComponent(txt_perfilEmpresa1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                .addGroup(painelFormLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(txt_perfilEmpresa7, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jLabel8))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                .addGroup(painelFormLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(jLabel7)
-                    .addComponent(txt_perfilEmpresa4, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addGap(18, 18, 18)
-                .addGroup(painelFormLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
-                    .addComponent(jLabel1)
-                    .addComponent(txt_perfilEmpresa5, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addGap(34, 34, 34)
-                .addGroup(painelFormLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(jLabel9)
-                    .addComponent(txt_perfilEmpresa8, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jLabel15)
-                    .addComponent(txt_perfilEmpresa12, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addGroup(painelFormLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(jLabel11)
-                    .addComponent(txt_perfilEmpresa11, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jLabel16)
-                    .addComponent(txt_perfilEmpresa15, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
-                .addGroup(painelFormLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(jLabel14)
-                    .addComponent(jTextField1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addGroup(painelFormLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(txt_perfilEmpresa13, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jLabel3)
-                    .addComponent(jLabel4)
-                    .addComponent(txt_perfilEmpresa16, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 41, Short.MAX_VALUE)
-                .addComponent(jButton2)
-                .addContainerGap())
-        );
-
-        javax.swing.GroupLayout painelBackgroundFormLayout = new javax.swing.GroupLayout(painelBackgroundForm);
-        painelBackgroundForm.setLayout(painelBackgroundFormLayout);
-        painelBackgroundFormLayout.setHorizontalGroup(
-            painelBackgroundFormLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, painelBackgroundFormLayout.createSequentialGroup()
-                .addContainerGap()
-                .addComponent(painelForm, javax.swing.GroupLayout.PREFERRED_SIZE, 550, Short.MAX_VALUE)
-                .addContainerGap())
-        );
-        painelBackgroundFormLayout.setVerticalGroup(
-            painelBackgroundFormLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(painelBackgroundFormLayout.createSequentialGroup()
-                .addContainerGap()
-                .addComponent(painelForm, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                .addContainerGap())
-        );
-
-        background.add(painelBackgroundForm, new org.netbeans.lib.awtextra.AbsoluteConstraints(230, 120, 570, 420));
-
-        javax.swing.GroupLayout jPanel1Layout = new javax.swing.GroupLayout(jPanel1);
-        jPanel1.setLayout(jPanel1Layout);
-        jPanel1Layout.setHorizontalGroup(
-            jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addComponent(background, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-        );
-        jPanel1Layout.setVerticalGroup(
-            jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel1Layout.createSequentialGroup()
-                .addComponent(background, javax.swing.GroupLayout.PREFERRED_SIZE, 551, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(0, 0, Short.MAX_VALUE))
-        );
-
-        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
-        getContentPane().setLayout(layout);
-        layout.setHorizontalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 810, Short.MAX_VALUE)
-            .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                .addGroup(layout.createSequentialGroup()
-                    .addGap(0, 0, Short.MAX_VALUE)
-                    .addComponent(jPanel1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addGap(0, 0, Short.MAX_VALUE)))
-        );
-        layout.setVerticalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGap(0, 551, Short.MAX_VALUE)
-            .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                .addGroup(layout.createSequentialGroup()
-                    .addGap(0, 0, Short.MAX_VALUE)
-                    .addComponent(jPanel1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addGap(0, 0, Short.MAX_VALUE)))
-        );
-
-        pack();
-    }// </editor-fold>//GEN-END:initComponents
-
-    private void txt_perfilEmpresa6ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_txt_perfilEmpresa6ActionPerformed
-        // TODO add your handling code here:
-    }//GEN-LAST:event_txt_perfilEmpresa6ActionPerformed
-
-    private void txt_perfilEmpresa5ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_txt_perfilEmpresa5ActionPerformed
-        // TODO add your handling code here:
-    }//GEN-LAST:event_txt_perfilEmpresa5ActionPerformed
-
-    /**
-     * @param args the command line arguments
-     */
-    public static void main(String args[]) {
-        /* Set the Nimbus look and feel */
-        //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
-        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
-         * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html 
-         */
-        try {
-            for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
-                if ("Nimbus".equals(info.getName())) {
-                    javax.swing.UIManager.setLookAndFeel(info.getClassName());
-                    break;
-                }
-            }
-        } catch (ClassNotFoundException ex) {
-            java.util.logging.Logger.getLogger(Cadastro_Cliente.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (InstantiationException ex) {
-            java.util.logging.Logger.getLogger(Cadastro_Cliente.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (IllegalAccessException ex) {
-            java.util.logging.Logger.getLogger(Cadastro_Cliente.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (javax.swing.UnsupportedLookAndFeelException ex) {
-            java.util.logging.Logger.getLogger(Cadastro_Cliente.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        }
-        //</editor-fold>
-
-        /* Create and display the form */
-        java.awt.EventQueue.invokeLater(new Runnable() {
-            public void run() {
-                new Cadastro_Cliente().setVisible(true);
-            }
-        });
-    }
-
-    // Variables declaration - do not modify//GEN-BEGIN:variables
-    private javax.swing.JPanel background;
-    private javax.swing.JPanel faixaTitulo;
-    private javax.swing.JLabel icon;
-    private javax.swing.JLabel icon1;
-    private javax.swing.JLabel icon2;
-    private javax.swing.JLabel icon3;
-    private javax.swing.JLabel iconLogo;
-    private javax.swing.JButton jButton2;
-    private javax.swing.JLabel jLabel1;
-    private javax.swing.JLabel jLabel11;
-    private javax.swing.JLabel jLabel13;
-    private javax.swing.JLabel jLabel14;
-    private javax.swing.JLabel jLabel15;
-    private javax.swing.JLabel jLabel16;
-    private javax.swing.JLabel jLabel3;
-    private javax.swing.JLabel jLabel4;
-    private javax.swing.JLabel jLabel5;
-    private javax.swing.JLabel jLabel6;
-    private javax.swing.JLabel jLabel7;
-    private javax.swing.JLabel jLabel8;
-    private javax.swing.JLabel jLabel9;
-    private javax.swing.JPanel jPanel1;
-    private javax.swing.JTextField jTextField1;
-    private javax.swing.JPanel linhaDivisoria;
-    private javax.swing.JPanel menuLateral;
-    private javax.swing.JLabel nome;
-    private javax.swing.JLabel nome1;
-    private javax.swing.JLabel nome2;
-    private javax.swing.JLabel nome3;
-    private javax.swing.JLabel nomePagina;
-    private javax.swing.JPanel opcaoMenu;
-    private javax.swing.JPanel opcaoMenu1;
-    private javax.swing.JPanel opcaoMenu2;
-    private javax.swing.JPanel opcaoMenu3;
-    private javax.swing.JPanel painelBackgroundForm;
-    private javax.swing.JPanel painelForm;
-    private javax.swing.JLabel subTitulo;
-    private javax.swing.JTextField txt_perfilEmpresa;
-    private javax.swing.JTextField txt_perfilEmpresa1;
-    private javax.swing.JTextField txt_perfilEmpresa11;
-    private javax.swing.JTextField txt_perfilEmpresa12;
-    private javax.swing.JTextField txt_perfilEmpresa13;
-    private javax.swing.JTextField txt_perfilEmpresa15;
-    private javax.swing.JTextField txt_perfilEmpresa16;
-    private javax.swing.JTextField txt_perfilEmpresa2;
-    private javax.swing.JTextField txt_perfilEmpresa4;
-    private javax.swing.JTextField txt_perfilEmpresa5;
-    private javax.swing.JTextField txt_perfilEmpresa6;
-    private javax.swing.JTextField txt_perfilEmpresa7;
-    private javax.swing.JTextField txt_perfilEmpresa8;
-    // End of variables declaration//GEN-END:variables
-}
